# aplicativos/backend/API/swagger.json
{
    "info": {
        "description": "This is a sample server",
        "title": "Swagger",
        "version": "1.0.0"
    },
    "host": "localhost:4000",
    "basePath": "/v1",
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http"
    ],
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "in": "header",
            "name": "Authorization",
            "description": "Bearer <token>"
        }
    },
    "swagger": "2.0",
    "paths": {
...
        "/security/login": {
            "post": {
                "parameters": [
                    {
                        "name": "post",
                        "in": "body",
                        "description": "the new user",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Login"
                        }
                    }
                ],
                "description": "\nThis function creates a user",
                "tags": [
                    "Security"
                ],
                "responses": {}
            }
        },
        "/security/register": {
            "post": {
                "parameters": [
                    {
                        "name": "post",
                        "in": "body",
                        "description": "the new user",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Registry"
                        }
                    }
                ],
                "description": "\nThis function creates a user",
                "tags": [
                    "Security"
                ],
                "responses": {}
            }
        },
        "/users/me": {
            "get": {
                "parameters": [],
                "description": "This function get my users",
                "tags": [
                    "User"
                ],
                "responses": {
                    "200": {
                        "description": "my user"
                    },
                    "default": {
                        "description": "Unexpected error"
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            },
            "put": {
                "parameters": [],
                "description": "This function update my user",
                "tags": [
                    "User"
                ],
                "responses": {
                    "200": {
                        "description": "my user"
                    },
                    "default": {
                        "description": "Unexpected error"
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            },
            "delete": {
                "parameters": [],
                "description": "This function update my user",
                "tags": [
                    "User"
                ],
                "responses": {
                    "200": {
                        "description": "my user"
                    },
                    "default": {
                        "description": "Unexpected error"
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        }
    },
    "definitions": {
...
        "Login": {
            "required": [
                "user",
                "password"
            ],
            "properties": {
                "user": {
                    "type": "string",
                    "description": ""
                },
                "password": {
                    "type": "string",
                    "description": ""
                }
            }
        },
        "Registry": {
            "required": [
                "user",
                "password"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "description": ""
                },
                "user": {
                    "type": "string",
                    "description": ""
                },
                "password": {
                    "type": "string",
                    "description": ""
                }
            }
        },
...
    },
...
}